flowers_api/requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
stripe==13.2.0
reportlab==4.4.4
openpyxl==3.1.5
celery==5.4.0
